==> Ruggeri et al (2020)/76/Analysis/code/02 Item-based analysis meta-analysis.ipynb <==
# Loading packages

import numpy as np
import pandas as pd
import glob

import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

from pandas.api.types import CategoricalDtype

# Loading data

data= pd.read_pickle("../data/modified_exclusions/pt_replication_modified_exclusions_data.pkl")

# Create original PT proportions

original_pt_proportions = pd.Series([0.18, 0.83, 0.20, 0.65, 0.14, 0.73, 0.92, 0.42, 0.92, 0.30, 0.22, 0.16, 0.69, 0.18, 0.70, 0.72, 0.17],
                                   index=[str(i) for i in np.arange(1, 18)])
original_pt_proportions

# Preparing long format data

long_data = data.groupby("Country")[[str(i) for i in np.arange(1, 18)]].mean().reset_index().melt(id_vars="Country", value_vars=[str(i) for i in np.arange(1, 18)],
                                                                                     var_name="Item", value_name="Proportion")

long_data2 = pd.DataFrame(original_pt_proportions)
long_data2 = long_data2.reset_index()
long_data2.columns = ["Item", "Proportion"]
long_data2["Item"] = long_data2["Item"].astype(str)
long_data2["Country"] = "Original"

# Data wrangling with the proportions

country_a_counts = data.groupby("Country")[[str(i) for i in np.arange(1, 18)]].sum()

country_sample_sizes = data.groupby("Country")[[str(i) for i in np.arange(1, 18)]].count()
country_b_counts = country_sample_sizes - country_a_counts

%load_ext RWinOut

%R -i country_a_counts
%R -i country_sample_sizes
%R -i country_b_counts

%%R
library(metafor)

%%R
rownames(country_a_counts)

%%R
fit = escalc(measure="PR", xi=country_a_counts$X1, ni=country_sample_sizes$X1, slab=rownames(country_a_counts))
summary(fit)

### Running a random-effects log-odds model on the item-by-item response proportion. I chose log-odds because they tend to be normally distributed (which improves the validity of the model inferences) and because 0 means that people were indifferent between the options, which is the main effect we want to investigate.

%%R
model_list = list()
estimate_list = list()
lb_list = list()
ub_list = list()
z_list = list()
p_list = list()
I2_list = list()
q_list = list()
qp_list = list()
X = 0
for (item in c("X1", "X2", "X3", "X4", "X5", "X6", "X7", "X8", "X9", "X10",
              "X11", "X12", "X13", "X14", "X15", "X16", "X17"))
{
X = X + 1
fit = rma.uni(measure="PLO", xi=country_a_counts[[item]], ni=country_sample_sizes[[item]],
              method="ML", slab=rownames(country_a_counts))
print(item)
print(summary(fit))
model_list[[X]] = fit
estimate_list[[X]] = fit$beta[[1]]
lb_list[[X]] = fit$ci.lb
ub_list[[X]] = fit$ci.ub
z_list[[X]] = fit$zval
p_list[[X]] = fit$pval
I2_list[[X]] = fit$I2
q_list[[X]] = fit$QM[[1]]
qp_list[[X]] = fit$QMp

}

%%R
library(gtools)

### Preparing tables for assessing heterogeneity and overall difference from chance

%%R
z_list = as.numeric(unlist(z_list))
p_list = as.numeric(unlist(p_list))
I2_list = as.numeric(unlist(I2_list))
q_list = as.numeric(unlist(q_list))
qp_list = as.numeric(unlist(qp_list))

%R -o z_list
%R -o p_list
%R -o I2_list
%R -o q_list
%R -o qp_list
z_list = np.array(z_list)
p_list = np.array(p_list)
I2_list = np.array(I2_list)
q_list = np.array(q_list)
qp_list = np.array(qp_list)

q_list

het_df = pd.DataFrame([[str(x) for x in np.arange(1, 18)], q_list, qp_list, I2_list, z_list, p_list]).transpose()
het_df.columns = ["Item", "Q (df = 18)", "Q p-value", "I2", "z-statistic", "z p-value"]
het_df["Q (df = 18)"] = het_df["Q (df = 18)"].astype(float).round(2)
het_df["I2"] = het_df["I2"].astype(float).round(0).astype(int)
het_df["z-statistic"] = het_df["z-statistic"].astype(float).round(2)

het_df.to_excel("../output/Meta analysis proportions heterogeneity.xlsx", index=False)
het_df.to_csv("../output/Meta analysis proportions heterogeneity.csv", index=False)

## Transforming estimates and upper and lower confidence bounds back to proportions

%%R
lb_list = sapply(as.numeric(unlist(lb_list)), inv.logit)
ub_list = sapply(as.numeric(unlist(ub_list)), inv.logit)
estimate_list = sapply(as.numeric(unlist(estimate_list)), inv.logit)
p_list = as.numeric(unlist(p_list))

%R -o lb_list
%R -o ub_list
%R -o estimate_list
%R -o p_list
lb_list = np.array(lb_list)
ub_list = np.array(ub_list)
estimate_list = np.array(estimate_list)
p_list = np.array(p_list)

prop_df = pd.DataFrame([estimate_list, lb_list, ub_list, p_list]).transpose()
prop_df.columns = ["prop", "lb", "ub", "p-value"]
prop_df.head()

prop_df["Item"] = [str(x) for x in np.arange(1, 18)]

prop_df["Succesful Replication"] = "Yes"
prop_df.loc[prop_df["p-value"]>.05, "Succesful Replication"] = "No"
prop_df.loc[prop_df["Item"]=="8", "Succesful Replication"] = "NA"

prop_df

prop_df.to_excel("../output/Meta analysis proportions.xlsx", index=False)
prop_df.to_csv("../output/Meta analysis proportions.csv", index=False)
