==> Ruggeri et al (2020)/76/Analysis/code/06 Bayesian Posterior Plots.ipynb <==
# Loading packages

import numpy as np
import pandas as pd
import glob

import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

from pandas.api.types import CategoricalDtype

# Loading data

import os
import glob

path = '../output/bayesian_item_posteriors/'
extension = 'csv'
os.chdir(path)
file_names = glob.glob('*.{}'.format(extension))
print(file_names)

raw_data = []
for file_name in file_names:
    raw_data.append(pd.read_csv(file_name, index_col=0))

path = "../../figures/supplement/bayesian_item_posteriors/"
for i in range(len(raw_data)):
    temp = raw_data[i]
    
    ### Sorting and labelling the columns
    old_col_names = temp.columns[:14]
    temp = temp.loc[:, old_col_names]
    new_col_names = [x[2:] for x in old_col_names]
    new_col_names[13] = "SD_Country_Intercept"
    new_col_names
    temp = temp.rename(columns=dict(zip(old_col_names, new_col_names)))
    temp = temp.melt(var_name="Coeffcient", value_name="Posterior")
    
    ### Plotting the figure
    sns.set(style="whitegrid", font_scale=1.25)
    fig, ax = plt.subplots(figsize=(8, 12))
    ax.axvline(ls='--', c="black", x=0, zorder=-1000)
    sns.violinplot(x="Posterior", y="Coeffcient", data=temp, ax = ax, color='#fdaa48',
              order=['Intercept',  'SD_Country_Intercept', 'Loss_Awareness',
     'Loss_Intuition', 'Men', 'Z_Age', 'Z_Income', 'Education.L', 'Education.Q', 'Education.C', 'Bill_Payments.L',
     'Bill_Payments.Q', 'Bill_Payments.C', 'Bill_PaymentsE4'], zorder=1000)
    ax.set(ylabel='', xlim=(-3, 3), title="Item " + str(i+1))
    
    ### Saving and closing the figure
    plt.savefig(path + "Item " + str(i+1) + ".svg", bbox_inches="tight")
    plt.close()
    print("Item " + str(i+1))
