==> Ruggeri et al (2020)/76/Analysis/code/07 Contrast pair analysis unpooled.ipynb <==
# Loading packages

import numpy as np
import pandas as pd
import glob

import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

from pandas.api.types import CategoricalDtype

# Loading data

data= pd.read_pickle("../data/modified_exclusions/pt_replication_modified_exclusions_data.pkl")

data.head()

# Testing for the key theoretical contrasts in the PT paper

PT_effects = {"Certainty Effect":[["1", "2"], ["3", "4"], ["7", "8"]],
            "Reflection Effect":[["3", "7"], ["4", "8"], ["5", "9"], ["6", "10"], ["16", "17"]],
             "Isolation Effect":[["4", "11"]],
             "Overweighting of small probabilities":[["5", "6"], ["9", "10"]],
             "Framing Effect":[["12", "13"]],
             "Range Adaptation":[["14", "15"]]}

from scipy.stats import fisher_exact

### Odds ratios for the original data

opt_table = pd.read_excel("../data/original_study_values.xlsx", index_col=0)

opt_table.head()

### Option A and B estimated are the number of observations in each cell based on the total N and proportions reported in the paper. Because observations by definition have to be integers it makes sense to round these estimates.

opt_table["Option A (estimated)"] = opt_table["Option A (estimated)"].round()
opt_table["Option B (estimated)"] = opt_table["Option B (estimated)"].round()

### Run Fisher's exact test on the original data, and add the results to the dataframe

%%time
PT_contrasts_original_df = pd.DataFrame(columns=["Effect", "Item 1", "Item 2", "A1", "B1", "A2", "B2", "OR", "p-value"])
for key in PT_effects.keys():
    effect = PT_effects[key] 
    for i in np.arange(len(effect)):
        q1 = effect[i][0]
        q2 = effect[i][1]
        a1 = opt_table.loc[int(q1), "Option A (estimated)"]
        b1 = opt_table.loc[int(q1), "Option B (estimated)"]
        a2 = opt_table.loc[int(q2), "Option A (estimated)"]
        b2 = opt_table.loc[int(q2), "Option B (estimated)"]
        odds_ratio, p_value = fisher_exact([[a1, b1], [a2, b2]])
        PT_contrasts_original_df.loc[len(PT_contrasts_original_df)] = [key, q1, q2, a1, b1, a2, b2, odds_ratio, p_value]
        

PT_contrasts_original_df["OR2"] = PT_contrasts_original_df["OR"]
PT_contrasts_original_df.loc[PT_contrasts_original_df["OR2"]<1, "OR2"] = 1/PT_contrasts_original_df.loc[PT_contrasts_original_df["OR2"]<1, "OR2"]
PT_contrasts_original_df["Significance"] = PT_contrasts_original_df["p-value"] < 0.05

## Odds ratio by country

### Run Fisher's exact test and add the results to the dataframe

%%time
PT_contrasts_df2 = pd.DataFrame(columns=["Country", "Effect", "Item 1", "Item 2", "A1", "B1", "A2", "B2", "OR", "p-value"])
for country in data["Country"].unique():
    df = data.loc[data["Country"]==country, :]
    for key in PT_effects.keys():
        effect = PT_effects[key] 
        for i in np.arange(len(effect)):
            q1 = effect[i][0]
            q2 = effect[i][1]
            a1 = df[q1].sum()
            b1 = df[q1].count()-df[q1].sum()
            a2 = df[q2].sum()
            b2 = df[q2].count()-df[q2].sum()
            odds_ratio, p_value = fisher_exact([[a1, b1], [a2, b2]])
            PT_contrasts_df2.loc[len(PT_contrasts_df2)] = [country, key, q1, q2, a1, b1, a2, b2, odds_ratio, p_value]
            
        

%%time
PT_contrasts_df2["OR2"] = PT_contrasts_df2["OR"]
PT_contrasts_df2.loc[PT_contrasts_df2["OR2"]<1, "OR2"] = 1/PT_contrasts_df2.loc[PT_contrasts_df2["OR2"]<1, "OR2"]
PT_contrasts_df2["Significance"] = PT_contrasts_df2["p-value"] < 0.05
PT_contrasts_df2["Contrast"] = [a + " vs " + b for a, b in zip(PT_contrasts_df2["Item 1"], PT_contrasts_df2["Item 2"])]

### Add the total sample size for each country to the dataframe

sample_sizes = data["Country"].value_counts()

PT_contrasts_df2["Sample Size"] = np.nan
for country in sample_sizes.index:
    PT_contrasts_df2.loc[PT_contrasts_df2["Country"]==country, "Sample Size"] = sample_sizes[country]


PT_contrasts_df2.head()

PT_contrasts_original_df.head()

### Transform significance from boolean to string

PT_contrasts_df2["Significant"] = np.nan
PT_contrasts_df2.loc[PT_contrasts_df2["Significance"]==True, "Significant"] = "Yes"
PT_contrasts_df2.loc[PT_contrasts_df2["Significance"]==False, "Significant"] = "No"

### Adding log odds

PT_contrasts_df2["LogOR"] = np.log(PT_contrasts_df2["OR"])

PT_contrasts_original_df["LogOR"] = np.log(PT_contrasts_original_df["OR"])

### Checking the extent to which the contrast effects replicate (given sign and signficance)

PT_contrasts_df2["Significance2"] = PT_contrasts_df2["Significance"].astype(float)
for ix in PT_contrasts_original_df.index:
    itm1 = PT_contrasts_original_df.loc[ix, "Item 1"]
    itm2 = PT_contrasts_original_df.loc[ix, "Item 2"]
    log_sign = np.sign(PT_contrasts_original_df.loc[ix, "LogOR"])
    PT_contrasts_df2.loc[(PT_contrasts_df2["Item 1"]==itm1) & (PT_contrasts_df2["Item 2"]==itm2) & (np.sign(PT_contrasts_df2["LogOR"])!=log_sign), "Significance2"] = 0
    

PT_contrasts_df2.groupby("Contrast")["Significance2"].mean()

PT_contrasts_df2.groupby("Effect")["Significance2"].mean()

PT_contrasts_df2["Significance2"].mean()

PT_contrasts_df2.groupby("Country")["Significance2"].mean().sort_values(ascending=False)

# Quantifying differences in contrast effects between the original data and the country samples

PT_contrasts_df2.head()

PT_contrasts_original_df.head()

### Compute the difference in log-odds between the original results and the country specific results transformed so that larger effects are positive and smaller effects are negative, irrespective of original sign.

PT_contrasts_df2["LogOR_difference"] = np.nan
for a, b in zip(PT_contrasts_original_df["Item 1"], PT_contrasts_original_df["Item 2"]): # Cycle through the items
    if np.sign(PT_contrasts_original_df.loc[(PT_contrasts_original_df["Item 1"]==a) & (PT_contrasts_original_df["Item 2"]==b),
                         "LogOR"]).values[0] == 1: # If the original log-odds are positive 
        PT_contrasts_df2.loc[(PT_contrasts_df2["Item 1"]==a) & (PT_contrasts_df2["Item 2"]==b),
                             "LogOR_difference"] =     PT_contrasts_df2.loc[
            (PT_contrasts_df2["Item 1"]==a) & (PT_contrasts_df2["Item 2"]==b),
                             "LogOR"] - PT_contrasts_original_df.loc[
            (PT_contrasts_original_df["Item 1"]==a) & (PT_contrasts_original_df["Item 2"]==b),
                             "LogOR"].values[0] # Compute the difference between country-effects and original
    else: # If the original log-odds are negative
            PT_contrasts_df2.loc[(PT_contrasts_df2["Item 1"]==a) & (PT_contrasts_df2["Item 2"]==b),
                             "LogOR_difference"] =     (PT_contrasts_df2.loc[
            (PT_contrasts_df2["Item 1"]==a) & (PT_contrasts_df2["Item 2"]==b),
                             "LogOR"] - PT_contrasts_original_df.loc[
            (PT_contrasts_original_df["Item 1"]==a) & (PT_contrasts_original_df["Item 2"]==b),
                             "LogOR"].values[0]) * -1 # Compute the difference between country-effects and original and multiply by negative 1

PT_contrasts_df2.head()

PT_contrasts_df2.groupby("Effect")["LogOR_difference"].median()

PT_contrasts_df2["Replicates"] = np.nan
PT_contrasts_df2.loc[PT_contrasts_df2["Significance2"]==1, "Replicates"] = "Yes"
PT_contrasts_df2.loc[PT_contrasts_df2["Significance2"]==0, "Replicates"] = "No"

PT_contrasts_df2.head()

display(PT_contrasts_df2["LogOR_difference"].mean())
display(PT_contrasts_df2["LogOR_difference"].std())

display(PT_contrasts_df2.groupby("Effect")["LogOR_difference"].mean())
display(PT_contrasts_df2.groupby("Effect")["LogOR_difference"].std())

### Saving country contrasts

PT_contrasts_df2.to_excel("../output/contrasts_by_country.xlsx", index=False)
PT_contrasts_df2.to_csv("../output/contrasts_by_country.csv", index=False)

### Saving original contrasts

PT_contrasts_original_df.to_excel("../output/original_contrasts.xlsx", index=False)
PT_contrasts_original_df.to_csv("../output/original_contrasts.csv", index=False)
