==> Ruggeri et al (2020)/76/Analysis/code/08 Contrast pair analysis meta-analysis.ipynb <==
# Loading packages

import numpy as np
import pandas as pd
import glob

import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

from pandas.api.types import CategoricalDtype

%load_ext RWinOut

%%R
library(gtools)
library(metafor)

# Loading data

### Odds ratio country by country

PT_contrasts_df2 = pd.read_excel("../output/contrasts_by_country.xlsx")

PT_contrasts_df2.head()

# Run meta-analytic models

%R -i PT_contrasts_df2

%%R
model_list = list()
estimate_list = list()
lb_list = list()
ub_list = list()
z_list = list()
p_list = list()
I2_list = list()
q_list = list()
qp_list = list()
X = 0
for (contrast in unique(PT_contrasts_df2$Contrast))
{
X = X + 1
ai = PT_contrasts_df2[PT_contrasts_df2["Contrast"]==contrast, "A1"]
bi = PT_contrasts_df2[PT_contrasts_df2["Contrast"]==contrast, "A2"]
ci = PT_contrasts_df2[PT_contrasts_df2["Contrast"]==contrast, "B1"]
di = PT_contrasts_df2[PT_contrasts_df2["Contrast"]==contrast, "B2"]
fit = rma.glmm(ai, bi, ci, di, measure="OR", method="ML", model="UM.RS")
print(contrast)
print(summary(fit))
model_list[[X]] = fit
estimate_list[[X]] = fit$beta[[1]]
lb_list[[X]] = fit$ci.lb
ub_list[[X]] = fit$ci.ub
z_list[[X]] = fit$zval
p_list[[X]] = fit$pval
I2_list[[X]] = fit$I2
q_list[[X]] = fit$QE.LRT[[1]]
qp_list[[X]] = fit$QEp.LRT

}



### Importing results from heterogeneity analysis to Python

%%R
z_list = as.numeric(unlist(z_list))
p_list = as.numeric(unlist(p_list))
I2_list = as.numeric(unlist(I2_list))
q_list = as.numeric(unlist(q_list))
qp_list = as.numeric(unlist(qp_list))

%R -o z_list
%R -o p_list
%R -o I2_list
%R -o q_list
%R -o qp_list
z_list = np.array(z_list)
p_list = np.array(p_list)
I2_list = np.array(I2_list)
q_list = np.array(q_list)
qp_list = np.array(qp_list)

### Structuring heterogeniety results into dataframe

het_df = pd.DataFrame([[str(a) + " vs " + str(b) for a, b in zip(PT_contrasts_df2.loc[PT_contrasts_df2["Country"]=="Sweden",
                                                    "Item 1"], PT_contrasts_df2.loc[PT_contrasts_df2["Country"]=="Sweden", "Item 2"])],
                       q_list, qp_list, I2_list, z_list, p_list]).transpose()
het_df.columns = ["Contrasts", "Q (df = 18)", "Q p-value", "I2", "z-statistic", "z p-value"]
het_df["Q (df = 18)"] = het_df["Q (df = 18)"].astype(float).round(2)
#het_df["Q p-value"] = het_df["Q p-value"].round(4)
het_df["I2"] = het_df["I2"].astype(float).round(0).astype(int)
het_df["z-statistic"] = het_df["z-statistic"].astype(float).round(2)
#het_df["z p-value"] = het_df["z p-value"].round(4)
het_df.head()
het_df.to_excel("../output/Meta analysis logodds heterogeneity.xlsx", index=False)
het_df.to_csv("../output/Meta analysis logodds heterogeneity.csv", index=False)

het_df

### Importing log-odds results to Python

%%R
lb_list = as.numeric(unlist(lb_list))
ub_list = as.numeric(unlist(ub_list))
estimate_list = as.numeric(unlist(estimate_list))
p_list = as.numeric(unlist(p_list))

%R -o lb_list
%R -o ub_list
%R -o estimate_list
%R -o p_list
lb_list = np.array(lb_list)
ub_list = np.array(ub_list)
estimate_list = np.array(estimate_list)
p_list = np.array(p_list)

### Structuring results into dataframe

lodds_df = pd.DataFrame([estimate_list, lb_list, ub_list, p_list]).transpose()
lodds_df.columns = ["lodds", "lb", "ub", "p-value"]
lodds_df.head()

lodds_df["Contrasts"] = [str(a) + " vs " + str(b) for a, b in zip(PT_contrasts_df2.loc[PT_contrasts_df2["Country"]=="Sweden",
                                                    "Item 1"], PT_contrasts_df2.loc[PT_contrasts_df2["Country"]=="Sweden", "Item 2"])]

lodds_df["Succesful Replication"] = "Yes"
lodds_df.loc[lodds_df["p-value"]>.05, "Succesful Replication"] = "No"

lodds_df

# Compute the difference in log-odds between the original results and the meta-analytic results transformed so that larger effects are positive and smaller effects are negative, irrespective of original sign.

PT_contrasts_original_df = pd.read_excel("../output/original_contrasts.xlsx")

PT_contrasts_original_df.tail()

### Split the contrast column in lor_df to two separate columns (one for each item)

lodds_df["Item 1"] = lodds_df["Contrasts"].str.split(" vs ", expand=True)[0]
lodds_df["Item 2"] = lodds_df["Contrasts"].str.split(" vs ", expand=True)[1]

lodds_df["Item 1"] = lodds_df["Item 1"].astype(int)
lodds_df["Item 2"] = lodds_df["Item 2"].astype(int)

### Compute the difference in effect size for the central estimates

lodds_df["LogOR_difference"] = np.nan
for a, b in zip(PT_contrasts_original_df["Item 1"], PT_contrasts_original_df["Item 2"]):
    if np.sign(PT_contrasts_original_df.loc[(PT_contrasts_original_df["Item 1"]==a) & (PT_contrasts_original_df["Item 2"]==b),
                         "LogOR"]).values[0] == 1: # If the original log-odds are positive
        lodds_df.loc[(lodds_df["Item 1"]==a) & (lodds_df["Item 2"]==b),
                             "LogOR_difference"] = lodds_df.loc[
            (lodds_df["Item 1"]==a) & (lodds_df["Item 2"]==b),
                             "lodds"] - PT_contrasts_original_df.loc[
            (PT_contrasts_original_df["Item 1"]==a) & (PT_contrasts_original_df["Item 2"]==b),
                             "LogOR"].values[0] # Compute the difference between meta analytic effects and original
    else: # If the original log-odds are negative
            lodds_df.loc[(lodds_df["Item 1"]==a) & (lodds_df["Item 2"]==b),
                             "LogOR_difference"] =     (lodds_df.loc[
            (lodds_df["Item 1"]==a) & (lodds_df["Item 2"]==b),
                             "lodds"] - PT_contrasts_original_df.loc[
            (PT_contrasts_original_df["Item 1"]==a) & (PT_contrasts_original_df["Item 2"]==b),
                             "LogOR"].values[0]) * -1 # Compute the difference between meta analytic effects and original and multiply by negative 1

### Compute the difference in effect size for the upper bounds

lodds_df["ub_difference"] = np.nan
for a, b in zip(PT_contrasts_original_df["Item 1"], PT_contrasts_original_df["Item 2"]):
    if np.sign(PT_contrasts_original_df.loc[(PT_contrasts_original_df["Item 1"]==a) & (PT_contrasts_original_df["Item 2"]==b),
                         "LogOR"]).values[0] == 1: # If the original log-odds are positive
        lodds_df.loc[(lodds_df["Item 1"]==a) & (lodds_df["Item 2"]==b),
                             "ub_difference"] =     lodds_df.loc[
            (lodds_df["Item 1"]==a) & (lodds_df["Item 2"]==b),
                             "ub"] - PT_contrasts_original_df.loc[
            (PT_contrasts_original_df["Item 1"]==a) & (PT_contrasts_original_df["Item 2"]==b),
                             "LogOR"].values[0] # Compute the difference between meta analytic effects and original
    else: # If the original log-odds are negative
            lodds_df.loc[(lodds_df["Item 1"]==a) & (lodds_df["Item 2"]==b),
                             "ub_difference"] =     (lodds_df.loc[
            (lodds_df["Item 1"]==a) & (lodds_df["Item 2"]==b),
                             "ub"] - PT_contrasts_original_df.loc[
            (PT_contrasts_original_df["Item 1"]==a) & (PT_contrasts_original_df["Item 2"]==b),
                             "LogOR"].values[0]) * -1 # Compute the difference between meta analytic effects and original and multiply by negative 1

### Compute the difference in effect size for the lower bounds

lodds_df["lb_difference"] = np.nan
for a, b in zip(PT_contrasts_original_df["Item 1"], PT_contrasts_original_df["Item 2"]):
    if np.sign(PT_contrasts_original_df.loc[(PT_contrasts_original_df["Item 1"]==a) & (PT_contrasts_original_df["Item 2"]==b),
                         "LogOR"]).values[0] == 1: # If the original log-odds are positive
        lodds_df.loc[(lodds_df["Item 1"]==a) & (lodds_df["Item 2"]==b),
                             "lb_difference"] =     lodds_df.loc[
            (lodds_df["Item 1"]==a) & (lodds_df["Item 2"]==b),
                             "lb"] - PT_contrasts_original_df.loc[
            (PT_contrasts_original_df["Item 1"]==a) & (PT_contrasts_original_df["Item 2"]==b),
                             "LogOR"].values[0] # Compute the difference between meta analytic effects and original
    else: # If the original log-odds are negative
            lodds_df.loc[(lodds_df["Item 1"]==a) & (lodds_df["Item 2"]==b),
                             "lb_difference"] =     (lodds_df.loc[
            (lodds_df["Item 1"]==a) & (lodds_df["Item 2"]==b),
                             "lb"] - PT_contrasts_original_df.loc[
            (PT_contrasts_original_df["Item 1"]==a) & (PT_contrasts_original_df["Item 2"]==b),
                             "LogOR"].values[0]) * -1 # Compute the difference between meta analytic effects and original and multiply by negative 1

### Assign effect names to the meta-analytic countrasts

for a, b in zip(PT_contrasts_original_df["Item 1"], PT_contrasts_original_df["Item 2"]):
    lodds_df.loc[(lodds_df["Item 1"]==a) & (lodds_df["Item 2"]==b),
                             "Effect"] =  PT_contrasts_original_df.loc[
            (PT_contrasts_original_df["Item 1"]==a) & (PT_contrasts_original_df["Item 2"]==b),
                             "Effect"].values[0]

lodds_df

lodds_df.to_excel("../output/Meta analysis log odds.xlsx", index=False)
lodds_df.to_csv("../output/Meta analysis log odds.csv", index=False)
