==> Ruggeri et al (2020)/76/Analysis/code/09 Individual PT Analyses.ipynb <==
# Loading packages

import numpy as np
import pandas as pd
import glob

import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

from pandas.api.types import CategoricalDtype

# Loading data

data= pd.read_pickle("../data/modified_exclusions/pt_replication_modified_exclusions_data.pkl")

data.head()

PT_contrasts_original_df = pd.read_excel("../output/original_contrasts.xlsx")
PT_contrasts_original_df.head()

### Quantifying how many people changed their mind in the original experiment

PT_contrasts_original_df["A_difference"] = PT_contrasts_original_df["A1"] - PT_contrasts_original_df["A2"]

PT_contrasts_original_df[["Item 1", "Item 2", "A_difference"]]

### Quantifying how many people changed their minds in the direction of Prospect Theory in the current study

contrast_cat = CategoricalDtype(categories=["EUT + PT", "PT",
                                            "Neither"],
                               ordered=True)
contrast_columns = []
for ix in PT_contrasts_original_df.index:
    i1 = PT_contrasts_original_df.loc[ix, "Item 1"]
    i2 = PT_contrasts_original_df.loc[ix, "Item 2"]
    i1 = str(i1)
    i2 = str(i2)
    diff = PT_contrasts_original_df.loc[ix, "A_difference"]
    column_title = "Contrast_" + i1 + '_' + i2
    data[column_title] = np.nan
    data.loc[(data[i1]==0) & (data[i2]==0), column_title] = "EUT + PT"
    data.loc[(data[i1]==1) & (data[i2]==1), column_title] = "EUT + PT"
    if diff > 0:
        data.loc[(data[i1]==1) & (data[i2]==0), column_title] = "PT"
        data.loc[(data[i1]==0) & (data[i2]==1), column_title] = "Neither"
    if diff < 0:
        data.loc[(data[i1]==0) & (data[i2]==1), column_title] = "PT"
        data.loc[(data[i1]==1) & (data[i2]==0), column_title] = "Neither"
    data[column_title] = data[column_title].astype(contrast_cat)
    contrast_columns.append(column_title)

data.columns

### Saving choice contrast data

data.to_excel("../output/pt_replication_modified_exclusions_data_CoM.xlsx", index=False)
data.to_pickle("../output/pt_replication_modified_exclusions_data_CoM.pkl")
data.to_csv("../output/pt_replication_modified_exclusions_data_CoM.csv", index=False)

### Displaying choice proportions per contrast

for column in contrast_columns:
    display(round(data[column].value_counts(sort=False, normalize=True), 2))

### Preparing a table for the figure

contrast_effect_df = data[contrast_columns].melt(var_name="Effect", value_name="Direction")
contrast_table = contrast_effect_df.pivot_table(index="Effect", columns="Direction", values="Direction", aggfunc=len)
contrast_table = (contrast_table.transpose()/contrast_table.sum(1)).transpose()*100
contrast_table = contrast_table.loc[contrast_columns[::-1], :]

### Contrast table summary statistics

contrast_table.mean()

contrast_table["PT"].max().round()

contrast_table["PT"].min().round()

contrast_table["Neither"].std()

contrast_table["PT"].std()

### Saving the contrast table

contrast_table.to_excel("../output/contrast_changes_of_mind.xlsx")
contrast_table.to_csv("../output/contrast_changes_of_mind.csv")
